chan_lcr: pull in asterisk.h early.
We need to include asterisk.h early to get declarations for __ast_calloc
and friends.

Not doing so will lead to compile errors due to redefined symbols, since
most other headers use them, but do not include the necessary header
that declares them, leading to implicitly defined function that will not
match the later declaration.

Should be compatible with Asterisk 1.2+.